api/requirements.txt
azure-functions>=1.17.0
azure-storage-blob>=12.19.0
fastapi>=0.104.1
python-dotenv>=1.0.0
python-multipart>=0.0.6
